refactor(bot): extract arbitrage execution into ArbExecutionMixin (phase 5)
Third and largest cut of the AuramaurBot split. The arb execution paths
(_execute_conditional / negrisk / internal / cross_exchange_arb), the arb
scanner + depth-research loops, and concentrated-position rebalancing move
verbatim into auramaur/bot_arb.py as ArbExecutionMixin. Methods still operate on
the bot's self; the callers that stay in bot.py (_task_trading_cycle,
_task_correlation_scan) reach them via self.

- An AST sweep of the moved block caught one symbol that the test paths don't
  exercise: `except BudgetExhausted` referenced an exception not imported in the
  new module (would NameError if that try raised) — now imported. The other
  flagged names (Database/AlertManager/RiskManager) are annotation-only (lazy)
  and the *_EVENT_PCT values are method-local; added the type-only imports under
  TYPE_CHECKING for cleanliness.

bot.py: 3232 -> 2218 lines (3819 -> 2218 across all three phase-5 cuts, -42%).
Parity proven by the existing arb / correlation / negrisk suites. Full suite:
837 passed.
